requirements.txt
mlflow==2.16.0
pydantic==2.8.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
PyYAML==6.0.2
